Update Drupal Coder module to latest dev version
The latest development version of the Coder module better reflects
the current state of Drupal Coding Standard than the current stable
release.

This commit also removes the dependency on codesniffer itself as
it is already a dependency of Coder. This has the added benefit
that we always use the latest version of codesniffer that is
compatible with Coder.

Output of changes performed by Composer in this commit:
Updating squizlabs/php_codesniffer (2.7.0 => 2.9.1)
Updating symfony/yaml (v2.8.26 => v2.8.27)
Removing drupal/coder (8.2.10)
Installing drupal/coder (dev-8.x-2.x 7588f40)