Add support for Blackfire profiling to the Open Social docker containers
This commit adds the blackfire docker image as a container in the
docker-compose file and links it to the web container. This makes
it possible to use the Chrome Blackfire plugin to analyse any
Open Social website.

For this commit to work properly it is required that the Blackfire
probe is added to the open_social_docker:dev image.

For more info on how to use Blackfire in Docker and to find out
what keys you should add to your .bash_profile go to
https://blackfire.io/docs/integrations/docker