Add docker-compose set-up for Real Time chat
This duplicates the `docker-compose.yml` as an alternative in case you
want to run Real-Time chat.

However we could also adjust it so that it could be composed on top of
the existing docker-compose.yml file. The biggest issue is that we
need to adjust `web` with a depends_on for RabbitMQ so that it waits for
RabbitMQ to be started. That's difficult for a composable file that's
not an override. I'm not aware of other solutions for optional
containers.